requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.35.92
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
sniffio==1.3.1
starlette==0.46.2
supabase==2.9.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
